Move into factory pattern for client implementations (#12)
This PR reworks the data publish clients' implementation
to use the factory pattern to unify their usage in the
main. The clients shall adhere to the DataPublisher protocol.

Stdout mode is now implemented as a DataPublisher client.

Syntax and type hint corrections in main. Added architecture
diagram with mermaid and reworked client importing.

Dropped support for Python 3.8 and 3.9.